+++ static_frame/core/index_base.py
@@ -115,63 +115,18 @@ class IndexBase(ContainerOperand):
 
         from static_frame import Index
         from static_frame import IndexGO
-        from static_frame import IndexHierarchy
-        from static_frame import IndexHierarchyGO
         from static_frame import IndexNanosecond
         from static_frame import IndexNanosecondGO
         from static_frame.core.index_datetime import IndexDatetime
 
-        if isinstance(value, pandas.MultiIndex):
-            if value.has_duplicates:
-                raise ErrorInitIndex(f'cannot create IndexHierarchy from a MultiIndex with duplicates: {value}')
-
-            value = value.remove_unused_levels()
-
-            # iterating over a hierarchical index will iterate over labels
-            name: tp.Optional[tp.Tuple[tp.Hashable, ...]] = tuple(value.names)
-
-            # if not assigned Pandas returns None for all components, which will raise issue if trying to unset this index.
-            if all(n is None for n in name): #type: ignore
-                name = None
-
-            hierarchy_constructor = IndexHierarchy if cls.STATIC else IndexHierarchyGO
-
-            def build_index(pd_idx: pandas.Index) -> Index:
-                # NOTE: Newer versions of pandas will not require Python date objects to live inside
-                # a DatetimeIndex. Instead, it will be a regular Index with dtype=object.
-                # Only numpy datetime objects are put into a DatetimeIndex.
-                if isinstance(pd_idx, pandas.DatetimeIndex):
-                    constructor: tp.Type[Index] = IndexNanosecond
-                else:
-                    constructor = Index
-
-                if cls.STATIC:
-                    return constructor(pd_idx, name=pd_idx.name)
-                return tp.cast(Index, constructor._MUTABLE_CONSTRUCTOR(pd_idx))
-
-            indices: tp.List[Index] = []
-            indexers: np.ndarray = np.empty((value.nlevels, len(value)), dtype=DTYPE_INT_DEFAULT)
-
-            for i, (levels, codes) in enumerate(zip(value.levels, value.codes)):
-                indexers[i] = codes
-                indices.append(build_index(levels))
-
-            indexers.flags.writeable = False
-
-            return hierarchy_constructor(
-                    indices=indices,
-                    indexers=indexers,
-                    name=name,
-                    )
-
-        elif isinstance(value, pandas.DatetimeIndex):
+        if isinstance(value, pandas.DatetimeIndex):
             # if IndexDatetime, use cls, else use IndexNanosecond
             if issubclass(cls, IndexDatetime):
                 return cls(value, name=value.name)
-            else:
-                if not cls.STATIC:
-                    return IndexNanosecondGO(value, name=value.name)
-                return IndexNanosecond(value, name=value.name)
+
+            if not cls.STATIC:
+                return IndexNanosecondGO(value, name=value.name)
+            return IndexNanosecond(value, name=value.name)
 
         if not cls.STATIC:
             return IndexGO(value, name=value.name)
